api/requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.2
Werkzeug>=2.2.0
python-dotenv>=1.0.0
mistralai>=0.0.6
psycopg2-binary>=2.9.5
youtube-transcript-api>=0.6.0
requests>=2.28.0
gunicorn>=20.1.0